+++ shotover-proxy/tests/cassandra_int_tests/mod.rs
@@ -1,3 +1,20 @@
+use crate::helpers::cassandra::{assert_query_result, execute_query, run_query, ResultValue};
+use crate::helpers::ShotoverManager;
+use cassandra_cpp::{stmt, Batch, BatchType, Error, ErrorKind};
+use cassandra_protocol::events::ServerEvent;
+use cassandra_protocol::frame::events::StatusChange;
+use cdrs_tokio::authenticators::StaticPasswordAuthenticatorProvider;
+use cdrs_tokio::cluster::session::{SessionBuilder, TcpSessionBuilder};
+use cdrs_tokio::cluster::NodeTcpConfigBuilder;
+use cdrs_tokio::load_balancing::RoundRobinLoadBalancingStrategy;
+use futures::future::{join_all, try_join_all};
+use rand::Rng;
+use serial_test::serial;
+use std::sync::Arc;
+use test_helpers::docker_compose::{run_command, DockerCompose};
+use tokio::time::timeout;
+use tokio::time::Duration;
+
 mod batch_statements;
 mod cache;
 mod collections;
@@ -341,4 +358,109 @@ async fn test_cassandra_request_throttling() {
     std::thread::sleep(std::time::Duration::from_secs(1)); // sleep to reset the window
 
     batch_statements::test(&connection);
+}
+
+#[tokio::test(flavor = "multi_thread")]
+#[serial]
+async fn test_events_keyspace() {
+    let mut rng = rand::thread_rng();
+
+    let _docker_compose = DockerCompose::new(
+        "tests/test-configs/cassandra-peers-rewrite/docker-compose-4.0-cassandra.yaml",
+    );
+
+    let _shotover_manager = ShotoverManager::from_topology_file(
+        "tests/test-configs/cassandra-peers-rewrite/topology.yaml",
+    );
+
+    let user = "cassandra";
+    let password = "cassandra";
+    let auth = StaticPasswordAuthenticatorProvider::new(&user, &password);
+    let config = NodeTcpConfigBuilder::new()
+        .with_contact_point("127.0.0.1:9044".into())
+        .with_authenticator_provider(Arc::new(auth))
+        .build()
+        .await
+        .unwrap();
+
+    let session = TcpSessionBuilder::new(RoundRobinLoadBalancingStrategy::new(), config).build();
+    let mut event_recv = session.create_event_receiver();
+
+    let mut tries = 0;
+    loop {
+        if tries > 3 {
+            panic!("did not receive the event after 3 tries.");
+        }
+
+        let n: u8 = rng.gen();
+        let create_ks = format!("CREATE KEYSPACE IF NOT EXISTS test_ks_{} WITH REPLICATION = {{ 'class' : 'SimpleStrategy', 'replication_factor' : 1 }};", n);
+
+        session
+            .query(create_ks)
+            .await
+            .expect("Keyspace creation error");
+
+        match timeout(Duration::from_secs(10), event_recv.recv()).await {
+            Ok(recvd) => {
+                if let Ok(event) = recvd {
+                    println!("{:?}", event);
+                    assert!(matches!(event, ServerEvent::SchemaChange { .. }));
+                    break;
+                };
+            }
+            Err(err) => {
+                println!("plus tries {}", err);
+                tries += 1;
+            }
+        }
+    }
+}
+
+#[tokio::test(flavor = "multi_thread")]
+#[serial]
+async fn test_events_node() {
+    let _docker_compose = DockerCompose::new(
+        "tests/test-configs/cassandra-peers-rewrite/docker-compose-4.0-cassandra.yaml",
+    );
+
+    let _shotover_manager = ShotoverManager::from_topology_file(
+        "tests/test-configs/cassandra-peers-rewrite/topology.yaml",
+    );
+
+    let user = "cassandra";
+    let password = "cassandra";
+    let auth = StaticPasswordAuthenticatorProvider::new(&user, &password);
+    let config = NodeTcpConfigBuilder::new()
+        .with_contact_point("127.0.0.1:9044".into())
+        .with_authenticator_provider(Arc::new(auth))
+        .build()
+        .await
+        .unwrap();
+
+    let session = TcpSessionBuilder::new(RoundRobinLoadBalancingStrategy::new(), config).build();
+
+    let mut event_recv = session.create_event_receiver();
+
+    run_command("docker", &["kill", "cassandra-three"]).unwrap();
+
+    let mut tries = 0;
+    loop {
+        if tries > 3 {
+            panic!("did not receive the event after 3 tries.");
+        }
+
+        match timeout(Duration::from_secs(10), event_recv.recv()).await {
+            Ok(recvd) => {
+                if let Ok(ServerEvent::StatusChange(StatusChange { addr, .. })) = recvd {
+                    assert_eq!(addr.addr.port(), 9044);
+                } else {
+                    panic!("expected ServerEvent::StatusChange, got {:?}", recvd);
+                }
+                break;
+            }
+            Err(_) => {
+                tries += 1;
+            }
+        }
+    }
 }
